pipeline/requirements.txt
pandas>=2.1
requests>=2.31
numpy>=1.26
scipy>=1.11
scikit-learn>=1.4
supabase>=2.3
python-dotenv>=1.0
pyyaml>=6.0
pydantic>=2.5
haversine>=2.8
tabulate>=0.9
rich>=13.7
